moved interactive part to another file
